group applicants services in a folder
