+++ app/controllers/applicants_organisations_controller.rb
@@ -43,7 +43,7 @@ class ApplicantsOrganisationsController < ApplicationController
   end
 
   def save_applicant
-    @save_applicant ||= SaveApplicant.call(
+    @save_applicant ||= Applicants::SaveApplicant.call(
       applicant: @applicant,
       organisation: @organisation,
       rdv_solidarites_session: rdv_solidarites_session
