+++ spec/services/applicants/save_applicant_spec.rb
@@ -1,4 +1,4 @@
-describe SaveApplicant, type: :service do
+describe Applicants::SaveApplicant, type: :service do
   subject do
     described_class.call(
       rdv_solidarites_session: rdv_solidarites_session,
